Search for modules in the current directory and in the std_lib directory.
The std_lib directory is found in the directory of the executable. This means
that the std_lib directory in the repository directory needs to be copied to
the build directory. Additionally, support was added for loading builtin
modules; ones that are written in c. Import searches for modules in the
following order; the current directory, std_lib and then builtin modules.