Add operator module and ability for objects to support operators (and calls).
The operator module exposes language operators as symbols. It also includes
a special symbol called 'call'. When an object uses a symbol as a key,
dispatching the symbol (or using the operator) on the object calls value with
the given arguments. Simply calling the object calls the value associated
with the call symbol.