Make keep values they were defined with
When a scope is created the values it refers to will not change even if
they are reassigned after the scope is created. However, if a value is
undefined when a scope is created, its value will change if the variable
is reassigned in the outer scope. This is so functions can reference
themselves and each other.