Make ThingType a class.
The only implementation is LegacyThingType, which calls fields for the
respective methods. This is so the old function calls (ie setThingTypeCall,
etc.) can set the methods of a class instance. Each ThingType will become
their own class and each instance will hold the custom data instead of having
a blob of memory following a pointer to a ThingType struct.

Also, runtime.hpp was renamed to runtime.h and the build script now uses g++
instead of gcc.